Publishing script: do not wrap passthrough copies
This makes it so we can write code for things like the `DIRegistry`
which works correctly at both authoring and publishing time: it means
the side-effect-type modules can be in "module mode" rather than purely
ambient mode in both contexts. It also updates the definition of the
`@ember/service/owner-ext.d.ts` to be in module mode.