Emit named items for built-in components
The `Input` and `Textarea` components were previously being emitted as
the direct result of the `opaquify` item, with no name. This made it
impossible for TS to *automatically* emit them (a) with a useful import
name or (b) with a nameable type. Introducing an intermediate binding
and a type defined as the `typeof` the named constant, the combination
of which becomes the default export, maintains the existing runtime
behavior exactly while making the types name-able.